MGDOBR-391: use scheduled clean up job for empty namespaces
