--- shard-operator/src/main/java/com/redhat/service/bridge/shard/operator/CleanUpService.java
@@ -1,6 +1,0 @@
-package com.redhat.service.bridge.shard.operator;
-
-public interface CleanUpService {
-
-    void cleanUpEmptyNamespaces();
-}
